fixed it

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@ int _atoi(char *s)
 {
 	int i, j, size;
 	int sign = 1;
-	long int n = 0;
+	int n = 0;
 	int flag;
 	int oc;
